+++ client/src/components/CategoryBadge/index.ts
@@ -1,8 +1,8 @@
+import './style.scss';
+
 import Component from 'src/lib/component';
 
 import { CATEGORY__INFO } from 'src/constant/category';
-
-import './style.scss';
 
 type CategoryType = {
   category: string;
