+++ client/src/lib/observer/type.ts
@@ -2,6 +2,7 @@ type stateModel = {
   _state: any;
   _observers: Set<() => void>;
 };
+
 export type globalStateType = {
   [key: string]: stateModel;
 };
