+++ client/src/pages/ChartPage/index.ts
@@ -1,16 +1,14 @@
+import './style.scss';
+
 import Component from 'src/lib/component';
 import { getState } from 'src/lib/observer';
 
 import CategoryStatistics from 'src/components/CategoryStatistics';
 import Trend from 'src/components/Trend';
+
 import { currentCategoryState } from 'src/store/statistics';
 
-import fetchWrapper from 'src/utils/fetchWrapper';
 import { objType } from 'src/type/type';
-
-import { STATISTICS_URL } from 'src/configs/urls';
-
-import './style.scss';
 
 interface StateType {
   current: string | null;
