+++ client/src/store/page.ts
@@ -1,4 +1,5 @@
 import { initState } from 'src/lib/observer';
+
 import MainPage from 'src/pages/MainPage';
 
 export const pageState = initState({
