+++ client/src/utils/dataSetting.ts
@@ -1,13 +1,13 @@
 import { getState, setState } from 'src/lib/observer';
 
-import { getTransaction } from 'src/api/transaction';
-import { getCalendarStatistics } from 'src/api/calendar';
-import { getChartStatistics } from 'src/api/chart';
-
 import { pageState } from 'src/store/page';
 import { transactionState } from 'src/store/transaction';
 import { calendarDataState } from 'src/store/calendar';
 import { statisticsState, CategoryStatisticsType } from 'src/store/statistics';
+
+import { getTransaction } from 'src/api/transaction';
+import { getCalendarStatistics } from 'src/api/calendar';
+import { getChartStatistics } from 'src/api/chart';
 
 //util안에 있을 애는 아닌 것 같은데 위치를 못잡겠음
